feat: create JWT strategy
This class validates the incoming bearer token with a given secret (just
like specified in the parent class constructor) under the hood

What we're doing here is to add additional logic for authentication:
1. checking if the user exists
2. checking if the user is active

Pretty straightforward. I like it.

This is the first time I will do this. If I want to create another
backend with Nest, I can just create a repository with this
authentication system and clone it